Int64 types substituted for Int type alias to allow 32bit compatibility

+++ src/PopStructures/constructors/age_ind.jl
@@ -1,14 +1,14 @@
 # base types for individuals in age-structured populations
 mutable struct Individual
-    age::Int64
+    age::Int
     genotype::Array{Float64,1}   # vector of genotypes
     phenotype::Array{Float64,1}  # vector of phenotypes
     fitness::Array{Float64,1}    # (survival, fertility)
 
     # constructors
     Individual() = new(0, Float64[], Float64[], [1.0, 1.0])
-    Individual(age::Int64) = new(age, Float64[], Float64[], [1.0, 1.0])
-    Individual(age::Int64, genotype::Vector{Float64}) = new(age, genotype, Float64[], [1.0, 1.0])
-    Individual(age::Int64, genotype::Vector{Float64}, phenotype::Vector{Float64}) = new(age, genotype, phenotype, [1.0, 1.0])
-    Individual(age::Int64, genotype::Vector{Float64}, phenotype::Vector{Float64}, fitness::Vector{Float64}) = new(age, genotype, phenotype, fitness)
+    Individual(age::Int) = new(age, Float64[], Float64[], [1.0, 1.0])
+    Individual(age::Int, genotype::Vector{Float64}) = new(age, genotype, Float64[], [1.0, 1.0])
+    Individual(age::Int, genotype::Vector{Float64}, phenotype::Vector{Float64}) = new(age, genotype, phenotype, [1.0, 1.0])
+    Individual(age::Int, genotype::Vector{Float64}, phenotype::Vector{Float64}, fitness::Vector{Float64}) = new(age, genotype, phenotype, fitness)
 end
